feat(cache): cache now obeys Age and a variety of other things (#13)
Fixes: #5

This patch uses [`http-cache-semantics`](https://npm.im/http-cache-semantics)
to provide more "complete" caching support. This means things like the Age
header are now obeyed, heuristics are calculated properly, CC:immutable is
revalidated, etc.

The patch is also a big mess, and m-f-h is due for a nice refactor. heh. :/ #18 is a note for that.